Reorganise routes and folder structure
