+++ upflux-client/src/features/PasswordSettings/PasswordSettingsRoute.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Navbar } from '../../features/Navbar/Navbar';
-import { PasswordSettingsContent } from '../../features/PasswordSettings/PasswordSettings';
-import { Footer } from '../../features/Footer/Footer';
+import { Navbar } from '../Navbar/Navbar';
+import { PasswordSettingsContent } from './PasswordSettings';
+import { Footer } from '../Footer/Footer';
 
 export const PasswordSettingsRoute: React.FC = () => {
   return (
