--- upflux-client/src/routes/Login/LoginRoute.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { LoginComponent } from '../../features/Login/Login';
-
-export const LoginRoute: React.FC = () => {
-  return <LoginComponent />;
-};
